+++ src/main/java/com/dfsek/paralithic/eval/parser/Parser.java
@@ -19,8 +19,8 @@ import com.dfsek.paralithic.functions.Function;
 import com.dfsek.paralithic.functions.dynamic.DynamicFunction;
 import com.dfsek.paralithic.functions.natives.NativeFunction;
 import com.dfsek.paralithic.functions.natives.NativeMath;
-import com.dfsek.paralithic.functions.operation.NodeFunction;
-import com.dfsek.paralithic.functions.operation.TernaryIfFunction;
+import com.dfsek.paralithic.functions.node.NodeFunction;
+import com.dfsek.paralithic.functions.node.TernaryIfFunction;
 import com.dfsek.paralithic.node.Constant;
 import com.dfsek.paralithic.node.special.InvocationVariableNode;
 import com.dfsek.paralithic.node.Node;
